requirements.txt
Flask~=3.0.2
gunicorn~=21.2.0
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
geopy==2.4.1
pydub==0.25.1
google-generativeai==0.4.1
SpeechRecognition==3.10.3